requirements.txt
aiosqlite==0.17.0
certifi==2021.10.8
charset-normalizer==2.0.12
docutils==0.18.1
idna==3.3
iso8601==1.0.2
Kivy==2.1.0
Kivy-Garden==0.1.5
kivymd==0.104.2
Pillow==9.1.0
Pygments==2.12.0
pypika-tortoise==0.1.4
pytz==2022.1
requests==2.27.1
tortoise-orm==0.19.0
typing-extensions==4.2.0
urllib3==1.26.9
